🚚 rename : CandidateBottomSheet 컴포넌트 domain 구조로 변경
domain 구조로 변경 완료
