+++ src/components/domain/meet/detail/index.ts
@@ -1,0 +1,1 @@
+export { default as CandidateBottomSheet } from '../detail/CandidateBottomSheet'
