+++ src/components/domain/meet/index.ts
@@ -1,0 +1,1 @@
+export * from './detail'
