+++ src/components/organisms/index.ts
@@ -8,7 +8,7 @@ export { default as CollaboList } from './CollaboList'
 export { default as MeetDeleteModal } from './MeetDeleteModal'
 export { default as MeetCreateForm } from './MeetCreateForm'
 export { default as RegisterList } from './RegisterList'
-export { default as CandidateBottomSheet } from './CandidateBottomSheet'
+export { default as CandidateBottomSheet } from '../domain/meet/detail/CandidateBottomSheet'
 export { default as CandidateCancelModal } from './CandidateCancelModal'
 export { default as FormHeader } from './FormHeader'
 export { default as PerformanceDetailSectionTop } from './PerformanceDetailSectionTop'
